Stop using turbopack
We're removing turbopack for now. We saw in production that turbopack was causing the preview container to crash when combined with Shadcn components. It's unclear why that was happening but it's likely related to how resource intensive turbopack can be. Removing it for now, we can evaluate bringing it back later.